requirements.txt
pandas
boto3
watchdog
kafka-python

#
pymysql